pyproject.toml
[tool.poetry]
name = "camserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.8.1.78"
transformers = "^4.35.2"
pillow = "^10.1.0"
torch = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
